+++ Assets/Scripts/Datas/CommonGameRules/CommonGameRulesAuthoring.cs
@@ -2,17 +2,17 @@
 using UnityEngine;
 
 public struct CommonGameRulesData : IComponentData {
-    public float rotateSpeed;
+    public int rotateSpeed;
 }
 
 public class CommonGameRulesAuthoring : MonoBehaviour {
-    public float rotateSpeed    = 1;
+    public int rotateSpeed = 1;
 
     private class Baker : Baker<CommonGameRulesAuthoring> {
         public override void Bake(CommonGameRulesAuthoring authoring) {
             var eneity = GetEntity(TransformUsageFlags.Dynamic);
             AddComponent(eneity, new CommonGameRulesData {
-                rotateSpeed    = authoring.rotateSpeed
+                rotateSpeed = authoring.rotateSpeed
             });
         }
     }
